📝 Added project idea to README
I've added a detailed description of the project idea in the README file. This includes the main functionality of searching YouTube with base64 encoded strings and an uploader function for videos with base64 encoded titles. Also, I mentioned the possibility of turning it into a Chrome extension that can be toggled on and off.